+++ app/Http/Controllers/JadwalController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\Auth;
 
 //model
 use App\Models\Guru;
@@ -21,25 +22,30 @@ class JadwalController extends Controller
      */
     public function index(string $id = null)
     {
-        //
-        $title = "Daftar Kelas";
-        // $jadwal = Jadwal::all();
-        $kelas = DB::table('kelas')
-        ->join('gurus', 'gurus.kelas_id', '=', 'kelas.id')
-        ->select('kelas.*', 'gurus.nama_guru')
-        ->get();
-        
-        
-        return view('dashboard.Jadwal.index',[
-            'title'=>$title,
-            'kelas'=>$kelas
+        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
+            //
+            $title = "Daftar Kelas";
+            // $jadwal = Jadwal::all();
+            $kelas = DB::table('kelas')
+                ->join('gurus', 'gurus.kelas_id', '=', 'kelas.id')
+                ->select('kelas.*', 'gurus.nama_guru')
+                ->orderBy('angka_kelas','asc')
+                ->get();
 
-            // 'kelas'=>$kelas,
-            // 'DataGuru'=>$DataGuru,
-            // 'mapel'=>$mapel,
-            // 'jadwal'=>$jadwal,
-            // 'DataGuru'=>$DataGuru,
-        ]);
+
+            return view('dashboard.Jadwal.index', [
+                'title' => $title,
+                'kelas' => $kelas
+
+                // 'kelas'=>$kelas,
+                // 'DataGuru'=>$DataGuru,
+                // 'mapel'=>$mapel,
+                // 'jadwal'=>$jadwal,
+                // 'DataGuru'=>$DataGuru,
+            ]);
+        } else {
+            return back();
+        }
     }
 
     /**
@@ -47,21 +53,27 @@ class JadwalController extends Controller
      */
     public function create()
     {
-        $title = "Input Data Jadwal Pelajaran";
-        // $DataGuru = Guru::all();
-        $DataGuru = DB::table('gurus')->join('kelas', 'gurus.kelas_id','=','kelas.id')->select('gurus.*','kelas.angka_kelas')->get();
+        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
 
-        $mapel = MataPelajaran::all();
-        $kelas = Kelas::all();
-        return view('dashboard.Jadwal.TambahDataJadwal',[
-            'title'=>$title,
-            'DataGuru'=>$DataGuru,
-            'mapel'=>$mapel,
-            'kelas'=>$kelas,
-        ]);
+            $title = "Input Data Jadwal Pelajaran";
+            // $DataGuru = Guru::all();
+            $DataGuru = DB::table('gurus')->join('kelas', 'gurus.kelas_id', '=', 'kelas.id')->select('gurus.*', 'kelas.angka_kelas', 'kelas.nama_kelas')->orderBy('angka_kelas','asc')->get();
+
+            $mapel = MataPelajaran::all();
+            $kelas = Kelas::all();
+            return view('dashboard.Jadwal.TambahDataJadwal', [
+                'title' => $title,
+                'DataGuru' => $DataGuru,
+                'mapel' => $mapel,
+                'kelas' => $kelas,
+            ]);
+        } else {
+            return back();
+        }
     }
 
-    public function tambahJadwal(string $id_kelas, $id_guru){
+    public function tambahJadwal(string $id_kelas, $id_guru)
+    {
 
     }
 
@@ -70,32 +82,34 @@ class JadwalController extends Controller
      */
     public function store(Request $request)
     {
-        $messages = [
-            'id_mapel.required' => 'Mata pelajaran harus dipilih.',
-            'id_guru.required' => 'Guru pengampu harus dipilih.',
-            'id_kelas.required' => 'Kelas harus dipilih.',
-            'jam_mulai.required' => 'Jam mulai harus diisi.',
-            'jam_selesai.required' => 'Jam selesai harus diisi.',
-            'jam_selesai.after' => 'Jam selesai harus setelah jam mulai.',
-            'jumlah_sesi.required' => 'Jumlah sesi harus diisi.',
-            'jumlah_sesi.min' => 'Jumlah sesi minimal adalah 1.',
-        ];
-        $validator = Validator::make($request->all(), [
-            'id_mapel'     => 'required',
-            'id_guru' => 'required|exists:gurus,id',
-            'id_kelas'     => 'required|exists:kelas,id',
-            'jam_mulai' => 'required|date_format:H:i',
-            'jam_selesai' => 'required|date_format:H:i|after:jam_mulai',
-            'jumlah_sesi' => 'required|min:1',
-            'hari' => 'required|',
-        ],$messages);
-        
-        if ($validator->fails()) {
-            return redirect()->back()->withErrors($validator)->withInput();
-        }
+        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
+
+            $messages = [
+                'id_mapel.required' => 'Mata pelajaran harus dipilih.',
+                'id_guru.required' => 'Guru pengampu harus dipilih.',
+                'id_kelas.required' => 'Kelas harus dipilih.',
+                'jam_mulai.required' => 'Jam mulai harus diisi.',
+                'jam_selesai.required' => 'Jam selesai harus diisi.',
+                'jam_selesai.after' => 'Jam selesai harus setelah jam mulai.',
+                'jumlah_sesi.required' => 'Jumlah sesi harus diisi.',
+                'jumlah_sesi.min' => 'Jumlah sesi minimal adalah 1.',
+            ];
+            $validator = Validator::make($request->all(), [
+                'id_mapel' => 'required',
+                'id_guru' => 'required|exists:gurus,id',
+                'id_kelas' => 'required|exists:kelas,id',
+                'jam_mulai' => 'required|date_format:H:i',
+                'jam_selesai' => 'required|date_format:H:i|after:jam_mulai',
+                'jumlah_sesi' => 'required|min:1',
+                'hari' => 'required|',
+            ], $messages);
+
+            if ($validator->fails()) {
+                return redirect()->back()->withErrors($validator)->withInput();
+            }
 
             // ELOQUENT
-            $jadwal = New Jadwal;
+            $jadwal = new Jadwal;
             $jadwal->id_mapel = $request->id_mapel;
             $jadwal->id_guru = $request->id_guru;
             $jadwal->id_kelas = $request->id_kelas;
@@ -105,18 +119,21 @@ class JadwalController extends Controller
             $jadwal->hari = $request->hari;
             $jadwal->save();
 
-        //  Jadwal::create([
-        //     'id_mapel'      => $id_mapel,
-        //     'id_guru' => $request->id_guru,
-        //     'id_kelas'     => $request->id_kelas,
-        //     'jam_mulai'     => $request->jam_mulai,
-        //     'jam_selesai'     => $request->jam_selesai,
-        //     'jumlah_sesi'     => $request->jumlah_sesi,
-        //     'hari'     => $request->hari,
-        // ]);
-        
-        return redirect()->route('jadwal.show',$request->id_kelas)->with('Success','Data berhasil ditambahkan');
- 
+            //  Jadwal::create([
+            //     'id_mapel'      => $id_mapel,
+            //     'id_guru' => $request->id_guru,
+            //     'id_kelas'     => $request->id_kelas,
+            //     'jam_mulai'     => $request->jam_mulai,
+            //     'jam_selesai'     => $request->jam_selesai,
+            //     'jumlah_sesi'     => $request->jumlah_sesi,
+            //     'hari'     => $request->hari,
+            // ]);
+
+            return redirect()->route('jadwal.show', $request->id_kelas)->with('Success', 'Data berhasil ditambahkan');
+        } else {
+            return back();
+        }
+
     }
 
     /**
@@ -124,35 +141,40 @@ class JadwalController extends Controller
      */
     public function show(string $id)
     {
-        // Mengatur locale Carbon ke bahasa Indonesia
-        Carbon::setLocale('id');
-        // Mengambil hari saat ini dalam bahasa Indonesia
-        $currentDay = Carbon::now()->translatedFormat('l');
-        $hariIni = strtolower($currentDay);
+        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
 
-            $title="Jadwal Kelas : ";
+            // Mengatur locale Carbon ke bahasa Indonesia
+            Carbon::setLocale('id');
+            // Mengambil hari saat ini dalam bahasa Indonesia
+            $currentDay = Carbon::now()->translatedFormat('l');
+            $hariIni = strtolower($currentDay);
+
+            $title = "Jadwal Kelas : ";
             $jadwal = Jadwal::all();
             $DataGuru = Guru::all();
             $mapel = MataPelajaran::all();
             $kelas = Kelas::all();
             $kelasNow = Kelas::find($id);
             $kelasAbs = $id;
-            $jadwal = DB::table('jadwals')->join('gurus','gurus.id','=','jadwals.id_guru')->join('kelas','kelas.id','=','jadwals.id_kelas')->join('mata_pelajarans','mata_pelajarans.id','=','jadwals.id_mapel')
-                ->select('jadwals.*','jadwals.id as id_jadwal','gurus.nama_guru','kelas.angka_kelas','kelas.id as id_kelas', 'mata_pelajarans.nama_pelajaran')->where('kelas.id',$id)
+            $jadwal = DB::table('jadwals')->join('gurus', 'gurus.id', '=', 'jadwals.id_guru')->join('kelas', 'kelas.id', '=', 'jadwals.id_kelas')->join('mata_pelajarans', 'mata_pelajarans.id', '=', 'jadwals.id_mapel')
+                ->select('jadwals.*', 'jadwals.id as id_jadwal', 'gurus.nama_guru', 'kelas.angka_kelas', 'kelas.id as id_kelas', 'mata_pelajarans.nama_pelajaran')->where('kelas.id', $id)
                 ->get();
-            
-                
 
-        return view('dashboard.Jadwal.DataJadwal',[
-            'title'=>$title,
-            'DataGuru'=>$DataGuru,
-            'mapel'=>$mapel,
-            'kelas'=>$kelas,
-            'kelasNow'=>$kelasNow,
-            'kelasAbs'=>$kelasAbs,
-            'jadwal'=>$jadwal,
-            'hariIni'=> $hariIni,
-        ]);
+
+
+            return view('dashboard.Jadwal.DataJadwal', [
+                'title' => $title,
+                'DataGuru' => $DataGuru,
+                'mapel' => $mapel,
+                'kelas' => $kelas,
+                'kelasNow' => $kelasNow,
+                'kelasAbs' => $kelasAbs,
+                'jadwal' => $jadwal,
+                'hariIni' => $hariIni,
+            ]);
+        } else {
+            return back();
+        }
     }
 
     /**
@@ -160,23 +182,29 @@ class JadwalController extends Controller
      */
     public function edit(string $id)
     {
-        $title="Edit Data Jadwal";
-        // $jadwal = Jadwal::all();
-        $DataGuru = Guru::all();
-        $mapel = MataPelajaran::all();
-        $kelas = Kelas::all();
-        $jadwal = DB::table('jadwals')->join('gurus','gurus.id','=','jadwals.id_guru')->join('kelas','kelas.id','=','jadwals.id_kelas')->join('mata_pelajarans','mata_pelajarans.id','=','jadwals.id_mapel')
-            ->select('jadwals.*','jadwals.id as id_jadwal','gurus.nama_guru','kelas.angka_kelas','kelas.id', 'mata_pelajarans.nama_pelajaran','mata_pelajarans.id as matapelajaran_id')->where('jadwals.id',$id)
-            ->first();
+        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
+
+            $title = "Edit jadwal";
+            // $jadwal = Jadwal::all();
+            $DataGuru = Guru::all();
+            $mapel = MataPelajaran::all();
+            $kelas = Kelas::all();
+            $jadwal = DB::table('jadwals')->join('gurus', 'gurus.id', '=', 'jadwals.id_guru')->join('kelas', 'kelas.id', '=', 'jadwals.id_kelas')->join('mata_pelajarans', 'mata_pelajarans.id', '=', 'jadwals.id_mapel')
+                ->select('jadwals.*', 'jadwals.id as id_jadwal', 'gurus.nama_guru', 'kelas.angka_kelas', 'kelas.id', 'mata_pelajarans.nama_pelajaran', 'mata_pelajarans.id as matapelajaran_id', 'kelas.nama_kelas')->where('jadwals.id', $id)
+                ->first();
 
             // dd($mapel);
-        return view('dashboard.Jadwal.DataEditJadwal',[
-            'title'=>$title,
-            'DataGuru'=>$DataGuru,
-            'mapel'=>$mapel,
-            'kelas'=>$kelas,
-            'jadwal'=>$jadwal, 
-        ]);
+            return view('dashboard.Jadwal.DataEditJadwal', [
+                'title' => $title,
+                'DataGuru' => $DataGuru,
+                'mapel' => $mapel,
+                'kelas' => $kelas,
+                'jadwal' => $jadwal,
+            ]);
+
+        } else {
+            return back();
+        }
     }
 
     /**
@@ -184,42 +212,47 @@ class JadwalController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $messages = [
-            'id_mapel.required' => 'Mata pelajaran harus dipilih.',
-            'id_guru.required' => 'Guru pengampu harus dipilih.',
-            'id_kelas.required' => 'Kelas harus dipilih.',
-            'jam_mulai.required' => 'Jam mulai harus diisi.',
-            'jam_selesai.required' => 'Jam selesai harus diisi.',
-            'jam_selesai.after' => 'Jam selesai harus setelah jam mulai.',
-            'jumlah_sesi.required' => 'Jumlah sesi harus diisi.',
-            'jumlah_sesi.min' => 'Jumlah sesi minimal adalah 1.',
-        ];
-        $validator = Validator::make($request->all(), [
-            'id_mapel'     => 'required',
-            'id_guru' => 'required|exists:gurus,id',
-            'id_kelas'     => 'required|exists:kelas,id',
-            'jam_mulai' => 'required|date_format:H:i',
-            'jam_selesai' => 'required|date_format:H:i|after:jam_mulai',
-            'jumlah_sesi' => 'required|min:1',
-            'hari' => 'required|',
-        ],$messages);
-        
-        if ($validator->fails()) {
-            return redirect()->back()->withErrors($validator)->withInput();
+        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
+
+            $messages = [
+                'id_mapel.required' => 'Mata pelajaran harus dipilih.',
+                'id_guru.required' => 'Guru pengampu harus dipilih.',
+                'id_kelas.required' => 'Kelas harus dipilih.',
+                'jam_mulai.required' => 'Jam mulai harus diisi.',
+                'jam_selesai.required' => 'Jam selesai harus diisi.',
+                'jam_selesai.after' => 'Jam selesai harus setelah jam mulai.',
+                'jumlah_sesi.required' => 'Jumlah sesi harus diisi.',
+                'jumlah_sesi.min' => 'Jumlah sesi minimal adalah 1.',
+            ];
+            $validator = Validator::make($request->all(), [
+                'id_mapel' => 'required',
+                'id_guru' => 'required|exists:gurus,id',
+                'id_kelas' => 'required|exists:kelas,id',
+                'jam_mulai' => 'required|date_format:H:i',
+                'jam_selesai' => 'required|date_format:H:i|after:jam_mulai',
+                'jumlah_sesi' => 'required|min:1',
+                'hari' => 'required|',
+            ], $messages);
+
+            if ($validator->fails()) {
+                return redirect()->back()->withErrors($validator)->withInput();
+            }
+
+            // Pastikan jadwal ditemukan
+
+            $jadwal = Jadwal::find($id);
+            $jadwal->id_mapel = $request->id_mapel;
+            $jadwal->id_guru = $request->id_guru;
+            $jadwal->id_kelas = $request->id_kelas;
+            $jadwal->jam_mulai = $request->jam_mulai;
+            $jadwal->jam_selesai = $request->jam_selesai;
+            $jadwal->jumlah_sesi = $request->jumlah_sesi;
+            $jadwal->hari = $request->hari;
+            $jadwal->save();
+            return redirect()->route('jadwal.show', $request->id_kelas)->with('Success', 'Data berhasil diubah');
+        } else {
+            return back();
         }
-
-          // Pastikan jadwal ditemukan
-
-        $jadwal = Jadwal::find($id);
-        $jadwal->id_mapel = $request->id_mapel;
-        $jadwal->id_guru = $request->id_guru;
-        $jadwal->id_kelas = $request->id_kelas;
-        $jadwal->jam_mulai = $request->jam_mulai;
-        $jadwal->jam_selesai = $request->jam_selesai;
-        $jadwal->jumlah_sesi = $request->jumlah_sesi;
-        $jadwal->hari = $request->hari;
-        $jadwal->save();
-        return redirect()->route('jadwal.show',$request->id_kelas)->with('Success','Data berhasil diubah');
 
     }
 
@@ -228,8 +261,12 @@ class JadwalController extends Controller
      */
     public function destroy(string $id)
     {
-        $jadwal = Jadwal::findOrFail($id);
-        $jadwal->delete();
-        return redirect()->route('jadwal.show',$id)->with('Success', 'Data berhasil dihapus.');
+        if (Auth::guard('guru')->user()->level == 'tata usaha' || Auth::guard('guru')->user()->level == 'wali kelas') {
+            $jadwal = Jadwal::findOrFail($id);
+            $jadwal->delete();
+            return redirect()->route('jadwal.show', $id)->with('Success', 'Data berhasil dihapus.');
+        } else {
+            return back();
+        }
     }
 }
